Change of method
I've decided that having a button on the page the plugin listens to to trigger installs is a bad idea for security reasons. Gamebanana is full of XSS exploits and it would be extremely fucking ugly if someone used a script to forcefully install UI themes into someone's browser. I've decided to simply embed the entire theme manager into the settings page. It will be less convenient but much safer. obthemebtn.js will be removed and replaced with a theme injector that just adds @imports, the bulk of the theme code will reside in the settings page itself once a save/load system is put in place.